Creation of the generator object which permit to easily add attributes and objects to daily events, stored as a MISP feed. Plus, script fromredis which pops queue element in redis to put them in the feed
